Windows port: moreh: fix MSVC narrowing errors
- moreh_adamw: static_cast size_t grid.y to uint32_t when building
  AdamwWorkSplit, and cast the uint32_t CB index loop variable to
  uint8_t in the CBFormatDescriptor aggregate init.
- moreh_group_norm: cast the uint32_t cb_index parameter to uint8_t in
  the CBFormatDescriptor aggregate init in push_cb_if_nonzero.

+++ ttnn/cpp/ttnn/operations/moreh/moreh_group_norm/device/moreh_group_norm_program_factory.cpp
@@ -41,7 +41,7 @@ static void push_cb_if_nonzero(
             .total_size = num_tiles * tile_size,
             .core_ranges = cores,
             .format_descriptors = {{CBFormatDescriptor{
-                .buffer_index = cb_index,
+                .buffer_index = static_cast<uint8_t>(cb_index),
                 .data_format = data_format,
                 .page_size = tile_size,
             }}},
